refactor: Phase 2 — extract findall engine and FieldCaptureSlices helper
Move findall scan logic from lib.rs into parser/findall_engine.rs and add
FieldCaptureSlices::from_parser to reduce duplicated field-slice construction.
