+++ formatparse-pyo3/src/lib.rs
@@ -244,145 +244,13 @@ fn findall(
         })
     });
     let parser = get_or_create_parser(pattern, extra_types_cloned)?;
-
-    // Fast path: if no custom converters and evaluate_result=True, use raw matching
-    // This defers all Python object creation until the end (batch conversion)
-    // CRITICAL: Do ALL regex matching OUTSIDE GIL, then batch convert inside GIL
-    let has_custom_converters = extra_types
-        .as_ref()
-        .map(|et| !et.is_empty())
-        .unwrap_or(false);
-    let has_nested_dicts = parser.has_nested_dict_fields.iter().any(|&b| b);
-    let has_nested_format_fields = parser
-        .field_specs
-        .iter()
-        .any(|s| matches!(s.field_type, FieldType::Nested));
-
-    if !has_custom_converters && evaluate_result && !has_nested_dicts && !has_nested_format_fields {
-        // Use raw matching path: collect all raw data first (NO GIL), then batch convert
-        let mut raw_results = Vec::new();
-        let search_regex = parser.get_search_regex(case_sensitive);
-        let mut last_end = 0;
-        let mut raw_path_failed = false;
-
-        // Collect all raw matches OUTSIDE GIL (no Python objects created yet)
-        // This is the key optimization: all CPU work happens without GIL
-        for cap_result in search_regex.captures_iter(string) {
-            let captures = cap_result.map_err(crate::error::fancy_regex_match_error)?;
-            let Some(full_match) = captures.get(0) else {
-                return Err(pyo3::exceptions::PyRuntimeError::new_err(
-                    "regex match missing capture group 0",
-                ));
-            };
-            let match_start = full_match.start();
-            let match_end = full_match.end();
-
-            if match_start < last_end {
-                continue;
-            }
-
-            // Try raw matching (no Python objects, no GIL needed)
-            match crate::parser::matching::match_with_captures_raw(
-                &captures,
-                string,
-                match_start,
-                &crate::parser::matching::FieldCaptureSlices {
-                    field_specs: &parser.field_specs,
-                    field_names: &parser.field_names,
-                    normalized_names: &parser.normalized_names,
-                    custom_type_groups: &parser.custom_type_groups,
-                    has_nested_dict_fields: &parser.has_nested_dict_fields,
-                    nested_parsers: &parser.nested_parsers,
-                },
-            ) {
-                Ok(Some(raw_data)) => {
-                    raw_results.push(raw_data);
-                    last_end = match_end;
-
-                    if match_start == match_end {
-                        last_end += 1;
-                    }
-                }
-                Ok(None) => {}
-                Err(_) => {
-                    // Custom / datetime / other types need the Python matcher; retry full scan.
-                    raw_path_failed = true;
-                    break;
-                }
-            }
-        }
-
-        if !raw_path_failed {
-            // Return Results object with raw data (lazy conversion)
-            // This avoids creating all ParseResult objects upfront
-            // The Results object is lightweight - just stores raw data
-            return Python::with_gil(|py| -> PyResult<PyObject> {
-                let results = Results::new(raw_results);
-                Py::new(py, results)?.into_py_any(py)
-            });
-        }
-    }
-
-    // Fallback: use Python path (for custom converters or evaluate_result=False)
-    // Optimized: Collect results first, then create PyList with items directly
-    Python::with_gil(|py| -> PyResult<PyObject> {
-        let search_regex = parser.get_search_regex(case_sensitive);
-        let mut results = Vec::new();
-        let mut last_end = 0;
-        let extra_types_for_matching = if let Some(ref et) = extra_types {
-            et
-        } else {
-            &HashMap::new()
-        };
-
-        for cap_result in search_regex.captures_iter(string) {
-            let captures = cap_result.map_err(crate::error::fancy_regex_match_error)?;
-            let Some(full_match) = captures.get(0) else {
-                return Err(pyo3::exceptions::PyRuntimeError::new_err(
-                    "regex match missing capture group 0",
-                ));
-            };
-            let match_start = full_match.start();
-            let match_end = full_match.end();
-
-            if match_start < last_end {
-                continue;
-            }
-
-            let extra_types_ref = &extra_types_for_matching;
-
-            if let Some(result) = crate::parser::matching::match_with_captures(
-                &captures,
-                &crate::parser::matching::CapturedMatchContext {
-                    pattern: &parser.pattern,
-                    fields: crate::parser::matching::FieldCaptureSlices {
-                        field_specs: &parser.field_specs,
-                        field_names: &parser.field_names,
-                        normalized_names: &parser.normalized_names,
-                        custom_type_groups: &parser.custom_type_groups,
-                        has_nested_dict_fields: &parser.has_nested_dict_fields,
-                        nested_parsers: &parser.nested_parsers,
-                    },
-                    py,
-                    custom_converters: extra_types_ref,
-                    evaluate_result,
-                },
-            )? {
-                results.push(result);
-                last_end = match_end;
-
-                if match_start == match_end {
-                    last_end += 1;
-                }
-            }
-        }
-
-        // Create PyList with items directly (more efficient than empty + append)
-        // Convert PyObject to Bound<PyAny> for PyList::new
-        let items: Vec<_> = results.iter().map(|obj| obj.bind(py)).collect();
-        let results_list = PyList::new(py, items)?;
-        Ok(results_list.into())
-    })
+    crate::parser::findall_engine::findall_matches(
+        parser,
+        string,
+        extra_types.as_ref(),
+        case_sensitive,
+        evaluate_result,
+    )
 }
 
 /// Iterator over non-overlapping matches (same scan as :func:`findall`, one item per step).
